+++ 3_Coding/TransitSys/common/src/main/java/edu/nju/umr/po/order/PaymentPO.java
@@ -1,6 +1,7 @@
 package edu.nju.umr.po.order;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Calendar;
 
 import edu.nju.umr.constants.DateFormat;
@@ -19,14 +20,14 @@ public class PaymentPO extends PO implements Serializable,KindGetter,OrderOper{
 	private String payer;
 	private String account;
 	private Pay kind;
-	private double amount;
+	private BigDecimal amount;
 	private String remarks;
 	private Calendar opTime;
 	private String opName;
 	private String userId;
 	
 	public PaymentPO(int id, Calendar date, String payer, String account,
-			Pay kind, double amount, String remarks, Calendar opTime,
+			Pay kind, BigDecimal amount, String remarks, Calendar opTime,
 			String opName,String userId) {
 		super();
 		this.id = id;
@@ -55,7 +56,7 @@ public class PaymentPO extends PO implements Serializable,KindGetter,OrderOper{
 	public Pay getKind() {
 		return kind;
 	}
-	public double getAmount() {
+	public BigDecimal getAmount() {
 		return amount;
 	}
 	public String getRemarks() {
